# openapi.yaml
openapi: 3.0.3
info:
  title: OpenClaw Revenue Engine
  version: 1.0.0
  description: Paid-lane classification, Stripe checkout, usage metering, and MCP tools.
servers:
  - url: http://localhost:3000
paths:
...
  /usage/events:
    post:
      summary: Record an idempotent usage event
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Event recorded
  /usage/summary:
    get:
      summary: Return the usage summary for one tenant
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Usage summary
        "400":
          description: tenant_id is required
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...